Organised init classes a bit
