+++ GitVersionCore/Configuration/Init/SetConfig/GlobalModeSetting.cs
@@ -1,6 +1,7 @@
-﻿namespace GitVersion.Configuration.Wizard.SetConfig
+﻿namespace GitVersion.Configuration.Init.SetConfig
 {
     using System.Collections.Generic;
+    using GitVersion.Configuration.Init.Wizard;
 
     public class GlobalModeSetting : ConfigInitWizardStep
     {
