+++ GitVersionCore/Configuration/Init/StepResult.cs
@@ -1,4 +1,4 @@
-namespace GitVersion
+namespace GitVersion.Configuration.Init
 {
     public class StepResult
     {
